Updated direct declarator to support returned suffix declarators

+++ include/types.h
@@ -23,7 +23,7 @@ typedef struct TypeNode {
   union {
   struct FunctionType { TypeNode *param_types; int param_type_count; int param_type_capacity; TypeNode *return_type; } function_type;
   struct PointerType { TypeNode *reference_type; } pointer_type;
-  struct ArrayType { TypeNode *element_type; int size; } array_type;
+  struct ArrayType { TypeNode *element_type; unsigned long size; } array_type;
   } data;
 } TypeNode;
